% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mySA-package.R
\docType{package}
\name{mySA-package}
\alias{mySA}
\alias{mySA-package}
\title{\packageTitle{mySA}}
\description{
\packageDescription{mySA}
}
\details{
The DESCRIPTION file:
\tabular{ll}{
Package: \tab mySA\cr
Type: \tab Package\cr
Version: \tab 0.2\cr
Date: \tab 2018-08-29\cr
Depends: \tab lhs,randtoolbox,doMC,parallel\cr
Encoding: \tab UTF-8\cr
License: \tab GPL (>= 3)\cr
LazyLoad: \tab yes\cr
}
}
\author{
\packageAuthor{mySA}

Maintainer: \packageMaintainer{mySA}
}
