% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_vbsa_new.R
\name{vbsa}
\alias{vbsa}
\title{Variance Based Sensitivity Analysis}
\usage{
vbsa(fn, ..., lower = -Inf, upper = Inf, N = 1000,
  rtype = c("quasirandom", "lhs"), scrambling = c("Owen",
  "Faure-Tezuka", "Owen+Faure-Tezuka", "none"), subsamp = NULL, Nb = 1,
  Nb.sig = 0.05, ncores = 1, verbose = TRUE, full.output = FALSE,
  na.handle = "stop", debug = FALSE)
}
\arguments{
\item{fn}{character with the name of a valid R function to be analised.
Its first argument MUST contain a K-dimensional variable representing the
input factors to be analised. Its output must either be a single value of
type numeric or an optionally named numeric vector. If the latter, sensitivity
indices will be calculated and output be compiled for each of the fn outputs.}

\item{...}{additional arguments for 'fn'}

\item{lower}{K-dimensional (optionally named) numeric vector with the minimum values for
each input factor.}

\item{upper}{K-dimensional (optionally named) numeric vector with the maximum values for
each input factor.}

\item{N}{number of input factor samples used for the low-discrepancy (Sobol')
or LHS design of the matrices A and B.}

\item{rtype}{character, indicating the type of (quasi-)random sampling design used
for matrices A and B. Valid values are: 'quasirandom' (quasi-random low-discrepancy
Sobol' sequences) and 'lhs' (latin hypercube sampling)}

\item{scrambling}{Shall scrambling be applied to the quasi-random sequences? 
Default value: 'none'. See \code{\link[randtoolbox]{sobol}}.}

\item{subsamp}{If given, sensitivity indices will be calculated for
(and optionally bootstraping be applied to) sub-samples of N for monitoring of
convergence. Either a vector of numeric values <= N or a single numeric value
resulting in a sequence of subsamples \code{seq(subsamp, N, by = subsamp)}.
NOTE: Values of one make no sense and will be disregarded with a warning.}

\item{Nb}{Number of bootstrapping samples to assess the robustness (i.e. a confidence
interval) of calculated sensitivity indices. Default: 1 (i.e. bootstrapping is
effectively not conducted). Note: This requires no additional evaluations of \code{fn}.}

\item{Nb.sig}{Significance level defining upper and lower bounds of the confidence
intervals of sensitivity indices. Default: 0.05. Only used if \code{Nb > 1}.}

\item{ncores}{\code{integer} specifying the number of CPU cores to be used.
If > 1, packages \code{\link[doMC]{doMC}} (Linux only!) and \code{\link[parallel]{parallel}}
are needed. Values > 1 only useful if \code{fun} is very complex and computationally
demanding, otherwise multiple thread handling will cause function slowdown! Default: 1.}

\item{verbose}{logical, should progress messages be printed out to the screen?}

\item{full.output}{logical, should the sampling matrices 'A' and 'B' along with all
the input factor sets and its corresponding outputs of 'fn' be included in the output
of this function? Default: \code{FALSE}.}

\item{na.handle}{character. How to handle \code{NA} and non-finite values returned by \code{fn}?
One of: \code{stop} - function will stop with an error (default); \code{remove} - problematic
values are removed with a warning (affects N, fn.counts, A, B, Ab, fn.A, fn.B, fn.Ab)
and the calculation of indices proceeded.}

\item{debug}{logical. If \code{TRUE}, the function's complete internal environment
will be saved for closer inspection. The following files will be written to \code{getwd()},
unless an error occurred beforehand: \code{vbsa_backup1.RData} after initialisation
and generation of sampling matrices; \code{vbsa_backup2.RData} after evaluation of
\code{fn}; \code{vbsa_backup3.RData} after checking (and possibliy alterating if
\code{na.handle = TRUE}) the output of \code{fn}; \code{vbsa_backup4.RData} after
calculation of sensitivity indices, before compiling the output. Default: \code{FALSE}.}
}
\value{
For each output of \code{fn} an optionally named (sub-)list with the following elements:

 *) N                      : number of input factor samples employed in the SA
 
 *) fn.counts              : total number of model evaluations
 
 *) Si                     : First order sensitivity indices for each of the K input factors (if \code{Nb > 1} this is the mean of the bootstrapping distribution)
 
 *) St                     : Total effects sensitivity indices for each of the K input factors (if \code{Nb > 1} this is the mean of the bootstrapping distribution)
 
 *) ranking                : Importance ranking each of the K input factors according to Si
 
 If \code{Nb > 1} in addition:
 
 *) Si.sd                  : Standard deviation of the Si bootrstapping distribution
 
 *) Si.lo                  : Lower bound of the confidence interval of Si (defined by significance level \code{Nb.sig})

 *) Si.up                  : Upper bound of the confidence interval of Si (defined by significance level \code{Nb.sig})
 
 *) St.sd                  : Standard deviation of the St bootrstapping distribution
 
 *) St.lo                  : Lower bound of the confidence interval of St (defined by significance level \code{Nb.sig})

 *) St.up                  : Upper bound of the confidence interval of St (defined by significance level \code{Nb.sig})
 
 *) ranking.boot           : As ranking above, but for each bootstrapping sample
   
 OPTIONAL (if \code{full.output = TRUE}):

 *) Matrix.A               : Sampling matrix A of dimension (N, K)
 
 *) Matrix.B               : Sampling matrix B of dimension (N, K)
 
 *) Matrix.Ab              : Re-sampled matrix Ab (radial sampling) of dimension (N*K, K)
 
 *) fn.A                   : N-dimensional vector of fn evaluations for Matrix A
 
 *) fn.B                   : N-dimensional vector of fn evaluations for Matrix B
 
 *) fn.Ab                  : Matrix of fn evaluations for Matrix Ab of dimension (N, K)
 
 If \code{Nb > 1} in addition:
 
 *) Si.boot                : Matrix of Si bootstrapping distribution of dimension (Nb, K)
 
 *) St.boot                : Matrix of St bootstrapping distribution of dimension (Nb, K)
 
 *) B                      : Bootstrap matrix (indices for resampling of fn.A, fn.B, and fn.Ab) of dimension (N, Nb)
 
 NOTE: if argument \code{subsamp} was given, each sensitivity index vector or matrix
 as well as matrix B will be extended by a dimension giving the results for each the sub-sample.
}
\description{
Variance based Sensitivity analysis for the input factors of a 
hydrological model following Saltelli et al., 2010.
}
\examples{
 if ( !require(randtoolbox) ) install.packages(randtoolbox)
 if ( !require(lhs) )         install.packages(lhs)

 nparam <- 3

 # Example: Ishigami test function 
 ishigami <- function(x, a=7, b=0.1) {
   sin(x[1]) + a*(sin(x[2]))^2 + b*(x[3]^4)*sin(x[1]) 
 }
 
 set.seed(123) # for reproducible results
 
 res <- vbsa(
   fn="ishigami", 
   rtype = "quasirandom",
   lower=rep(-pi, nparam), 
   upper=rep(pi, nparam),
   N=2000, scrambling="none"
 )
 
}
\references{
Theory:
 
 Andrea Saltelli, Paola Annoni, Ivano Azzini, Francesca Campolongo,
 Marco Ratto, Stefano Tarantola, Variance based sensitivity
 analysis of model output. Design and estimator for the total
 sensitivity index, Computer Physics Communications, Volume 181,
 Issue 2, February 2010, Pages 259-270, DOI: 10.1016/j.cpc.2009.09.018.
 \url{http://www.sciencedirect.com/science/article/pii/S0010465509003087}
 
 Code based on R-script by Mauricio Zambrano-Bigiarini obtained from
 \url{https://ec.europa.eu/jrc/en/samo/simlab} (version from 20-May-2013),
 based on Matlab code by Stefano Tarantola
}
\author{
Tobias Pilz \email{tpilz@uni-potsdam.de}
}
