% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_vbsa.R
\name{vbsa_old}
\alias{vbsa_old}
\title{Variance Based Sensitivity Analysis}
\usage{
vbsa_old(fn, ..., lower = -Inf, upper = Inf, N = 1000,
  rtype = c("quasirandom", "lhs"), scrambling = c("Owen",
  "Faure-Tezuka", "Owen+Faure-Tezuka", "none"), reltol = 0, ncores = 1,
  verbose = TRUE, REPORT = 10, full.output = FALSE,
  monitor.ind = FALSE)
}
\arguments{
\item{fn}{character with the name of a valid R function to be analised.
Its first argument MUST contain a K-dimensional variable representing the
input factors to be analised.}

\item{...}{additional arguments for 'fn'}

\item{lower}{K-dimensional (optionally named) numeric vector with the minimum values for
each input factor.}

\item{upper}{K-dimensional (optionally named) numeric vector with the maximum values for
each input factor.}

\item{N}{number of input factor samples used for the low-discrepancy (Sobol')
or LHS design of the matrices A and B.}

\item{rtype}{character, indicating the type of (quasi-)random sampling design used
for matrices A and B. Valid values are: 'quasirandom' (quasi-random low-discrepancy
Sobol' sequences) and 'lhs' (latin hypercube sampling)}

\item{scrambling}{Shall scrambling be applied to the quasi-random sequences? 
Default value: 'none'. See \code{\link[randtoolbox]{sobol}}.}

\item{reltol}{numeric, relative tolerance. If a value different from zero,
the SA will finish when the maximum difference between the first
order sensitivity indices of all the input factors ('DSi') satisfy
the following relation:  Di = Max {|S[i] - S[i-1]} <= reltol}

\item{ncores}{\code{integer} specifying the number of CPU cores to be used.
If > 1, packages \code{\link[doMC]{doMC}} (Linux only!) and \code{\link[parallel]{parallel}}
are needed. Values > 1 only useful if \code{fun} is very complex and computationally
demanding, otherwise multiple thread handling will cause function slowdown! Default: 1.}

\item{verbose}{logical, should progress messages be printed out to the screen?}

\item{REPORT}{numeric, only used when 'verbose=TRUE'. A progress message is
printed every 'REPORT' number of model evaluations.}

\item{full.output}{logical, should the sampling matrices 'A' and 'B' along with all
the input factor sets and its corresponding outputs of 'fn' be included in the output
of this function? Default: \code{FALSE}.}

\item{monitor.ind}{logical, shall sensitivity indices Si and St be monitored?
If \code{FALSE} (default), output elements \code{First.Order.Indices.Si} and
\code{Total.Order.Indices.Sti} are merely K-dimensional numerical vectors considering
the full sample size. If \code{TRUE}, they are N-by-K dimensional matrices.}
}
\value{
A list with the following elements:

 *) N                      : number of input factor samples employed in the SA
 
 *) counts                 : total number of model evaluations
 
 *) First.Order.Indices.Si : First order sensitivity indices for each of the K input factors
 
 *) Sum.Si                 : sum of all the K first order sensitivity indices Si
 
 *) Total.Order.Indices.Sti: Total effects sensitivity indices for each of the K input factors
 
 *) Total.Indices.of.Pairs.of.Factors: matrix with the total indices of pairs of factors
 for each possible combination of input factors
 
 *) Ranking                : matrix with a ranking of the most important input factors,
 sorted in decreasing order according to the first order sensitivity index Si

 *) Matrix.A               : (optional). A matrix used in the radial sampling
 
 *) Matrix.B               : (optional). B matrix used in the radial sampling
 
 *) ParameterSets          : (optional) matrix with all the parameter sets used in the analysis
 
 *) GoFs                   : (optional) numeric vector with the outputs of 'fn' for each
 input factor set in 'ParameterSets'
}
\description{
Variance based Sensitivity analysis for the input factors of a 
hydrological model following Saltelli et al., 2010.
}
\examples{
 if ( !require(randtoolbox) ) install.packages(randtoolbox)
 if ( !require(lhs) )         install.packages(lhs)

 nparam <- 3

 # Example: Ishigami test function 
 ishigami <- function(x, a=7, b=0.1) {
   sin(x[1]) + a*(sin(x[2]))^2 + b*(x[3]^4)*sin(x[1]) 
 }
 
 set.seed(123) # for reproducible results
 
 res <- vbsa(
   fn="ishigami", 
   rtype = "quasirandom",
   lower=rep(-pi, nparam), 
   upper=rep(pi, nparam),
   N=2000, scrambling="none", REPORT=100
 )
 
}
\references{
Theory:
 
 Andrea Saltelli, Paola Annoni, Ivano Azzini, Francesca Campolongo,
 Marco Ratto, Stefano Tarantola, Variance based sensitivity
 analysis of model output. Design and estimator for the total
 sensitivity index, Computer Physics Communications, Volume 181,
 Issue 2, February 2010, Pages 259-270, DOI: 10.1016/j.cpc.2009.09.018.
 \url{http://www.sciencedirect.com/science/article/pii/S0010465509003087}
 
 Code based on R-script by Mauricio Zambrano-Bigiarini obtained from
 \url{https://ec.europa.eu/jrc/en/samo/simlab} (version from 20-May-2013),
 based on Matlab code by Stefano Tarantola
}
\author{
Tobias Pilz \email{tpilz@uni-potsdam.de}
}
